backend/pyproject.toml
[project]
name = "ailearning-backend"
version = "0.1.0"
description = "AI Learning Backend API"
readme = "../README.md"
requires-python = ">=3.12"
dependencies = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "langchain[openai]>=0.3.25",
    "langchain-openai>=0.3.17",
    "openai-agents[litellm]>=0.0.15",
    "python-dotenv>=1.1.0",
    "reportlab>=4.4.1",
    "langchain-community>=0.3.24",
    "langgraph>=0.4.5",
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "sqlalchemy[asyncio]>=2.0.41",
    "asyncpg>=0.29.0",
    "alembic>=1.13.1",
    "python-dotenv>=1.1.0",
    "pydantic[email]>=2.6.1",
    "pydantic-settings>=2.1.0",
    "aioredis>=2.0.1",
    "redis>=6.1.0",
    "arq>=0.25.0",
    "broadcaster>=0.3.1",
    "pyjwt>=2.10.1",
    "passlib>=1.7.4",
    "fastapi-security>=0.1.0",
    "email-validator>=2.2.0",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.alembic]
script_location = "app/migrations"
